fix: rework date state
