+++ src/components/CurrentTime.tsx
@@ -1,9 +1,16 @@
-import { useTime } from "@/contexts/TimeContext/useTime";
 import { formatDate } from "@/utils/common/formatters";
 import { FunctionComponent, useEffect, useState } from "react";
 
 export const CurrentTime: FunctionComponent = () => {
-  const { currentTime } = useTime();
+  const [currentTime, setCurrentTime] = useState<Date | null>(null);
+
+  useEffect(() => {
+    const intervalId = setInterval(() => {
+      setCurrentTime(new Date());
+    }, 1000);
+
+    return () => clearInterval(intervalId);
+  }, []);
 
   return (
     <div>
