+++ src/hooks/useLogPersonDetails.tsx
@@ -1,18 +1,15 @@
 import { useLogs } from "@/contexts/LogsContext/useLogs";
-import { useTime } from "@/contexts/TimeContext/useTime";
+import { formatDate } from "@/utils/common/formatters";
 import { User } from "@/utils/common/person";
 import { useEffect, useRef } from "react";
 
 export const useLogPersonData = (personData: User | undefined): void => {
   const { enableLogs } = useLogs();
-  const { currentTime } = useTime();
-
-  const time = useRef<Date | null>(null);
-
-  time.current = currentTime;
 
   useEffect(() => {
     if (enableLogs && personData) {
+      const time = formatDate(new Date());
+
       console.log("Person details:", personData);
       console.log("Current time:", time);
     }
