+++ src/pages/_app.tsx
@@ -1,5 +1,4 @@
 import { LogsContextProvider } from "@/contexts/LogsContext/LogsContext";
-import { TimeContextProvider } from "@/contexts/TimeContext/TimeContext";
 import "@/styles/globals.css";
 import type { AppProps } from "next/app";
 import { QueryClient, QueryClientProvider } from "react-query";
@@ -10,9 +9,7 @@ export default function App({ Component, pageProps }: AppProps) {
   return (
     <QueryClientProvider client={queryClient}>
       <LogsContextProvider>
-        <TimeContextProvider>
-          <Component {...pageProps} />
-        </TimeContextProvider>
+        <Component {...pageProps} />
       </LogsContextProvider>
     </QueryClientProvider>
   );
